Formato.

+++ src/caching/BaseTwoLevelStringFileCache.php
@@ -4,7 +4,6 @@ namespace minga\framework\caching;
 
 use minga\framework\Context;
 use minga\framework\IO;
-use minga\framework\Params;
 use minga\framework\Profiling;
 use minga\framework\settings\CacheSettings;
 
@@ -29,16 +28,13 @@ class BaseTwoLevelStringFileCache
 		$key2 = (string)$key2;
 
 		$folder = $this->path . "/" . $key1;
-		if ($key2 === '')
+		if ($key2 === '' && file_exists($folder))
 		{
-			if (file_exists($folder))
-			{
-				if (is_dir($folder))
-					IO::RemoveDirectory($folder);
-				else
-					IO::Delete($folder);
-				return;
-			}
+			if (is_dir($folder))
+				IO::RemoveDirectory($folder);
+			else
+				IO::Delete($folder);
+			return;
 		}
 		$file = $this->ResolveFilename($key1, $key2, false);
 		IO::Delete($file);
@@ -49,9 +45,8 @@ class BaseTwoLevelStringFileCache
 	public function HasData($key1, $key2, &$value = null) : bool
 	{
 		if (Context::Settings()->Cache()->Enabled !== CacheSettings::Enabled)
-		{
 			return false;
-		}
+
 		$file = $this->ResolveFilename($key1, $key2);
 		if (file_exists($file))
 		{
@@ -68,9 +63,8 @@ class BaseTwoLevelStringFileCache
 	public function HasRawData($key1, $key2, &$value = null) : bool
 	{
 		if (Context::Settings()->Cache()->Enabled !== CacheSettings::Enabled)
-		{
 			return false;
-		}
+
 		$file = $this->ResolveFilenameRaw($key1, $key2);
 		if (file_exists($file))
 		{
@@ -94,11 +88,10 @@ class BaseTwoLevelStringFileCache
 				IO::EnsureExists($folder);
 			return $folder . "/" . $key2 . ".txt";
 		}
-		else
-			return $this->path . "/" . $key1 . ".txt";
+
+		return $this->path . "/" . $key1 . ".txt";
 	}
 
-	
 	private function ResolveFilenameRaw($key1, $key2, bool $create = false) : string
 	{
 		$key1 = (string)$key1;
@@ -110,8 +103,8 @@ class BaseTwoLevelStringFileCache
 				IO::EnsureExists($folder);
 			return $folder . "/" . $key2 . ".raw";
 		}
-		else 
-			return $this->path . "/" . $key1 . ".raw";
+
+		return $this->path . "/" . $key1 . ".raw";
 	}
 
 	public function PutDataIfMissing($key1, $key2, $value) : void
